Day 1 Parts 1 and 2.
Part one is a very simple 'you can load a file and add it up, right?'

Part two is still pretty straightforward, loop over the input tracking
the current frequency and if we've seen it (HashSet.insert returns false
when inserting a duplicate) return that.

Adds in the test_case crate, for terse table-based testing.