Day 3 Part 1: Count overlapping claims
Simple use of `ndarray` to simulate the 1000x1000 fabric, then increment
each square in the claim, and count where claims > 1.

Also bring in `regex`, `lazy_static` because parsing.

`anyhow` for simple/lazy error handling in the Claim FromStr trait.  We
don't really need to wrap up things cleanly, so this is a little shorter
than making an error enum for `thiserror`