Day 4 Part 2: Identifty the guard with the maximal sleeping minute
Extracts exploit_weakest_guard function that takes the event stream and
a sort function to sort guards.

In part 1 the sort function is as before, total minutes slept.

In part 2 the sort function is the highest number of
times a guard was asleep in any minute (max(Guard.minutes.values)).