Day 5 Part 1 Two ways
The same algorithm twice, but in the original case we do a lot more
reallocation as we eagerly remove the two elements when we find them.

Doing it efficiently with retain needs a little more bookkeeping -
introduce a vector to track the elements we want to delete for, then
use while to simulate a C-Style for loop so we can advance over the
character if we match it, finally make the reduction.

  AOC 2018
  Day 5 - Part 1 - original : 11590
          generator: 3.592µs,
          runner: 629.408441ms

  Day 5 - Part 1 - retain : 11590
          generator: 326ns,
          runner: 18.003898ms