Day 07 Part 02: Lexical Topological Scheduler
Take the path generator from before with some small changes.

Track the working elves, and when they'll be done in the working
hashmap.  Fill the working set until all elves are fully tasked, or
there's nothing left to schedule (our answer == that time).
Advance time to the next tick where elves will be done and do
housekeeping on the working set.