Day 8 Part 1: Proof of tree parsing
Took a little thinking to figure out the simplest way to handle parsing
the tree out.   Got a little hung up initially on slicing off metadata
from the end, but that wouldn't work when there are multiple children at
a nested level.  Instead think more like nom, return the Node, and the
offset where it ended, then we can use that size value to handle knowing
where the end of the child we're processing is, without overrunning its
siblings.

Puzzle is then a simple tree walk over that to sum the values.