Day 12 Part 2: Detect when the function goes linear
Looking at the numbers we can see that after a while the step -> score
relationship goes linear.

Here we keep a window of 5 scores in a VecDeque, and keep the most
recent diffs in it.  Once all the diffs are the same we've hit the
linear part, and can simply calculate what it will be after 50 billion
steps.