Day 16 Part 2: Decode the instructions and run
It turns out that there's lots of information available in the checks
section.  Each pass over the checks we find there's one check that
matches exactly one opcode, so we can record that in a hashmap and loop
until it's got all 16 opcodes mapped to indexes.   Then evaluation is a
simple fold.