Day 18 Part 2: Run for 1 billion ticks
Well actually, we track the previous states and when they occurred so we
can find the point at which they recur, then advance time to just before
the 1 billionth second and resume ticking.

This is pretty much the same as we needed to do for 2017 Day 16 Part 1,
but we don't loop through the start state.

https://github.com/richardc/advent-go/blob/main/2017/day16/day16.go#L55-L70