Day 21: WatchASM returns
Extract the WASM (WatchASM) cpu to its own module so we can use it in
both day 19 and day 21 solutions.

For day 21, we have to analyse the code to see this at the end:

eqrr 4 0 2
addr 2 1 1
seti 5 8 1

Our pc register is 1 so that's just checking if register 0 (our minimal
value) matches register 4 (the solution for this part) are the same.

If so, we set pc to out of bounds - reg 2 will be 1, so addr 2 1 1 is a
jump over the next seti instuction and into space/halting.

Given this, our solution just runs until the pc is 3 before the end (the
eqrr), and then just returns the a register from that instruction.