Day 24 Part 2: Find minimal winning boost
This is a simple linear scan, we could pribably make it faster with a
binary search, but it'll do for now.

Main fiddliness is that in some boosted circumstances the game can
deadlock, so we needed to change play_to_win to return an Option, with
None in case of stalemate.